# openapi.yaml
openapi: 3.0.0
paths:
...
  "/comment/{id}/reply":
    post:
      operationId: replyOnComment
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
info:
  title: My API
  description: An API to do awesome things
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...